+++ resources/views/livewire/editorjs.blade.php
@@ -1,3 +1,9 @@
-<div x-data="editorInstance('data', '{{ $editorId }}', {{ $readOnly ? 'true' : 'false' }})" x-init="init()" class="{{ $class }}" style="{{ $style }}" wire:ignore>
+<div
+    x-data="editorInstance('data', '{{ $editorId }}', {{ $readOnly ? 'true' : 'false' }}, '{{ $placeholder }}', '{{ $logLevel }}')"
+    x-init="init()"
+    class="{{ $class }}"
+    style="{{ $style }}"
+    wire:ignore
+>
     <div id="{{ $editorId }}"></div>
 </div>