+++ src/Http/Livewire/EditorJS.php
@@ -23,18 +23,23 @@ class EditorJS extends Component
 
     public $readOnly;
 
+    public $placeholder;
+
     public $uploadDisk;
 
     public $downloadDisk;
+
+    public $logLevel;
 
     public function mount(
         $editorId,
         $value = [],
         $class = '',
         $style = '',
+        $readOnly = false,
+        $placeholder = null,
         $uploadDisk = null,
-        $downloadDisk = null,
-        $readOnly = false
+        $downloadDisk = null
     ) {
         if (is_null($uploadDisk)) {
             $uploadDisk = config('livewire-editorjs.default_img_upload_disk');
@@ -44,13 +49,20 @@ class EditorJS extends Component
             $downloadDisk = config('livewire-editorjs.default_img_download_disk');
         }
 
+        if (is_null($placeholder)) {
+            $placeholder = config('livewire-editorjs.default_placeholder');
+        }
+
         $this->editorId = $editorId;
         $this->data = $value;
         $this->class = $class;
         $this->style = $style;
+        $this->readOnly = $readOnly;
+        $this->placeholder = $placeholder;
         $this->uploadDisk = $uploadDisk;
         $this->downloadDisk = $downloadDisk;
-        $this->readOnly = $readOnly;
+
+        $this->logLevel = config('livewire-editorjs.editorjs_log_level');
     }
 
     public function completedImageUpload(string $uploadedFileName, string $eventName, $fileName = null)