+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to `livewire-editorjs` will be documented in this file
 
+## 1.0.0 2021-01-08
+
+- Added option to disable the component of the package
+- Added config option to rename the component of the package
+- Added `placeholder` to the component
+- Added `style` attribute to the component
+- Added `logLevel` to the Editor.js config
+- Added make command to create Editor.js components
+- Added test suite
+
+## 0.4.0 2021-01-04
+
+- Added `readOnly` mode to the Editor.
+- Lowered required Laravel version (7 & 8 supported)
+
 ## 0.3.0 - 2021-01-03
 
 Renamed method name for image upload. 
@@ -9,8 +24,8 @@ Like this it is more consistent when the `attaches` plugin of Editor.Js will be 
 
 ## 0.2.0 - 2021-01-03
 
-- Use `editorId` in the view as container id for the Editor.js instance
+Use `editorId` in the view as container id for the Editor.js instance
 
 ## 0.1.0 - 2021-01-03
 
-- Initial release of the package
+Initial release of the package